Rename SendToRow to AddRecipient
