+++ ui/app/pages/send/send-content/add-recipient/index.js
@@ -1,0 +1,1 @@
+export { default } from './add-recipient.container'
