+++ ui/app/pages/send/send-content/send-content.component.js
@@ -5,7 +5,7 @@ import SendAmountRow from './send-amount-row'
 import SendFromRow from './send-from-row'
 import SendGasRow from './send-gas-row'
 import SendHexDataRow from './send-hex-data-row'
-import SendToRow from './send-to-row'
+import SendToRow from './add-recipient'
 import SendAssetRow from './send-asset-row'
 
 export default class SendContent extends Component {
