--- ui/app/pages/send/send-content/send-to-row/index.js
@@ -1,1 +1,0 @@
-export { default } from './send-to-row.container'
