+++ ui/app/pages/send/send-content/tests/send-content-component.test.js
@@ -7,7 +7,7 @@ import PageContainerContent from '../../../../components/ui/page-container/page-
 import SendAmountRow from '../send-amount-row/send-amount-row.container'
 import SendFromRow from '../send-from-row/send-from-row.container'
 import SendGasRow from '../send-gas-row/send-gas-row.container'
-import SendToRow from '../send-to-row/send-to-row.container'
+import SendToRow from '../add-recipient/add-recipient.container'
 import SendHexDataRow from '../send-hex-data-row/send-hex-data-row.container'
 import SendAssetRow from '../send-asset-row/send-asset-row.container'
